Footer and move to spareted ViewComponent

+++ Custom_FrameWork/Application/Tools.cs
@@ -101,7 +101,7 @@ namespace Custom_FrameWork.Application
             return new DateTime(year, month, day, new PersianCalendar());
         }
 
-        public static string ToMoney(this double myMoney)
+        public static string ToMoney(this decimal myMoney)
         {
             return myMoney.ToString("N0", CultureInfo.CreateSpecificCulture("fa-ir"));
         }
